requirements.txt
werkzeug
python-dotenv
tweepy
flask